+++ elasticsearch_dsl/__init__.py
@@ -2,13 +2,13 @@ from .query import Q
 from .filter import F
 from .aggs import A
 from .function import SF
-from .search import Search
+from .search import Search, MultiSearch
 from .field import *
 from .document import DocType, MetaField
 from .mapping import Mapping
 from .index import Index
 from .analysis import analyzer, token_filter, char_filter, tokenizer
-from .faceted_search import * 
+from .faceted_search import *
 
 VERSION = (0, 0, 9)
 __version__ = VERSION