+++ Changelog.rst
@@ -9,6 +9,7 @@ Changelog
  * ``Search`` now caches responses from Elasticsearch
  * ``DateHistogramFacet`` now defaults to returning empty intervals
  * ``Search`` no longer accepts positional parameters
+ * Experimental ``MultiSearch`` API
 
 0.0.9 (2015-10-26)
 ------------------